requirements.txt
Flask==3.0.0
line-bot-sdk==3.9.0
gunicorn==21.2.0
